+++ package.json
@@ -26,6 +26,7 @@
 		"express": "^4.18.2",
 		"joi": "^17.6.3",
 		"pg": "^8.8.0",
+		"url-metadata": "^2.5.0",
 		"uuid": "^9.0.0"
 	},
 	"devDependencies": {
