--- src/db.js
@@ -1,14 +1,0 @@
-import pg from "pg";
-
-const { Pool } = pg;
-
-const databaseConfig = {
-    connectionString: process.env.DATABASE_URL,
-    ssl: {
-        rejectUnauthorized: false
-    }
-};
-
-const connection = new Pool (databaseConfig);
-
-export default connection;
