--- src/routes/postRoute.js
@@ -1,9 +1,0 @@
-import express from "express";
-import { listPosts } from "../controller/postControllers.js";
-import { authMiddleware } from "../middleware/authMiddleware.js"
-
-const router = express.Router()
-
-router.get("/posts", listPosts)
-
-export default router
