+++ src/routes/postsRoute.js
@@ -1,9 +1,10 @@
 import express from 'express';
-import * as postsController from '../controllers/postsController.js';
+import { sendPost, listPosts } from '../controllers/postsController.js';
 import { authMiddleware } from '../middleware/authMiddleware.js';
 
 const router = express.Router();
 
-router.post('/posts/publish', authMiddleware, postsController.sendPost);
+router.post('/posts/publish', authMiddleware, sendPost);
+router.get('/posts', listPosts);
 
 export default router;
